+++ packages/afterburner/common/lib.ts
@@ -1,3 +1,7 @@
 import { Construct } from '@aws-cdk/core';
 
 export type Builder<T, S = Construct> = (scope: S, id: string) => T;
+
+export function builderOf<T>(thing: T): Builder<T> {
+  return (scope: Construct, id: string) => thing;
+}