+++ packages/jet-nx/src/generators/versions.ts
@@ -1,0 +1,2 @@
+export const jet = '^0.0.4';
+export const afterburner = '^0.0.3';